refactor(regression): extract shared slope inference helper from OLS and TLS
Move the slope stderr / p-value / intercept_stderr formula into a shared
`calculate_slope_inference(n, ss_effective, ss_res, slope, x_mean)`
function in `utils/statistics.rs`.

OLS passes `ss_effective = Sxx`; TLS passes `ss_effective = Sxx*`.
Both paths now apply the same robustness guards (finite-sd check,
safe_divide, NaN fallbacks) that were previously only in OLS.

The now-redundant private `calculate_standard_errors` in `ols.rs` is
deleted; the `calculate_p_value_exact` re-export that no internal
non-test code required is removed from `utils/mod.rs`.
