+++ src/regression/utils/mod.rs
@@ -6,5 +6,5 @@ pub mod validation;
 
 // Re-export commonly used functions
 pub use math::kahan_sum;
-pub use statistics::{calculate_p_value_exact, compute_r_value};
+pub use statistics::{calculate_slope_inference, compute_r_value};
 pub use validation::{safe_divide, validate_finite_array};
